backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.48
Werkzeug==2.3.7
python-dotenv==1.0.0
openai==0.28.1
bcrypt==4.0.1
marshmallow==3.20.1
requests==2.31.0
plotly==5.17.0
networkx==3.2.1
matplotlib==3.8.2
seaborn==0.13.0
Pillow==10.1.0
pandas==2.1.4
kaleido==0.2.1